rename the project name into `petroly`
