+++ Procfile
@@ -1,2 +1,2 @@
 release: python manage.py migrate --noinput
-web: gunicorn kfupm.wsgi --log-file -
+web: gunicorn petroly.wsgi --log-file -
